+++ permissions/src/epilot/models/operations/assignroles.py
@@ -3,16 +3,14 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from typing import Optional
-
+from typing import List, Optional
 
 
 @dataclasses.dataclass
 class AssignRolesRequest:
     user_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'userId', 'style': 'simple', 'explode': False }})
-    request_body: Optional[list[str]] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})
+    request_body: Optional[List[str]] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})
     
-
 
 
 
@@ -22,7 +20,7 @@ class AssignRolesResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    assignments: Optional[list[str]] = dataclasses.field(default=None)
+    assignments: Optional[List[str]] = dataclasses.field(default=None)
     r"""ok"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""