+++ permissions/src/epilot/models/operations/getassignedrolesforuser.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from typing import Optional
-
+from typing import List, Optional
 
 
 @dataclasses.dataclass
@@ -14,14 +13,13 @@ class GetAssignedRolesForUserRequest:
 
 
 
-
 @dataclasses.dataclass
 class GetAssignedRolesForUserResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    assignments: Optional[list[str]] = dataclasses.field(default=None)
+    assignments: Optional[List[str]] = dataclasses.field(default=None)
     r"""ok"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""