+++ permissions/src/epilot/models/operations/refreshpermissions.py
@@ -6,7 +6,6 @@ import requests as requests_http
 from typing import Optional
 
 
-
 @dataclasses.dataclass
 class RefreshPermissionsResponse:
     content_type: str = dataclasses.field()